moving formatDate and public/js scripts
