--- public/js/endpoint.js
@@ -1,12 +1,0 @@
-function formatDate(dateString) {
-    const date = new Date(dateString);
-    const month = date.getUTCMonth() + 1; // Months are zero-based
-    const day = date.getUTCDate();
-    const year = date.getUTCFullYear();
-    const hours = date.getUTCHours().toString().padStart(2, '0');
-    const minutes = date.getUTCMinutes().toString().padStart(2, '0');
-    const seconds = date.getUTCSeconds().toString().padStart(2, '0');
-
-    return `${month}/${day}/${year}, ${hours}:${minutes}:${seconds}`;
-}
-module.exports = { formatDate };
